Move away from using integer-encoded policies.
Instead, this patch moves to generating one-hot policies for every
legal move in a single step and then using hashable references to
those policy tensors as keys to store the child states that result.

This results in significantly fewer expensive tf.scatter_nd calls
to generate one-hot policies, and because the policies can be reused
it also results in fewer calls to generate them overall. The result
is a significant performance improvement on both CPU and GPU platforms.

Signed-off-by: Peter <peter@psanders.me>