Restructure to comply with src layout and use ruff
We change to a single top-level import, "mcts", in order to support a
development workflow based on an editable install, migrating to a
pyproject.toml file instead of the old setup.py. This brings the project
up-to-date with PEP621, PEP631, and related.

We also add several more lints and other checks to pre-commit, and fix
the complaints of those checkers in order to baseline the repository.