Add an initial sketch of scalable gym code
The architecture for horizontal scalability comes from appendix D of
the MuZero paper [1]. We have a shared storage to which we write data
and models. Each data generator worker shall start a training cycle by
loading the latest model and running self-play, writing the replay
buffers back to shared storage. Meanwhile, the model-training worker
shall routinely rebuild the model using the latest data from the
workers.

The functionality is not yet complete in this commit, however this
module is imported by the agent code as committed, this submodule is
included for completeness to ensure all parts can be imported.

1: https://arxiv.org/pdf/1911.08265.pdf