Add non-working code for multi-process training
This code has known issues: particularly, mcts.storage._storage.Storage
writes empty files to disk, so improvements are needed to the storage
code. As follow-on work, I will be developing unit tests to ensure the
behavior of the Storage class so that datasets are correctly generated
and read.

With an alternate implementation of this Storage module (or with a
shared network file system), this code can be used to scale data
generation across arbitrarily many machines so that training datasets
can be more efficiently generated. This is especially valuable for games
that require a greater depth of planning (i.e., more nodes evaluated per
move) and for games that are computationally expensive to play compared
to the optimized pure tensorflow tic-tac-toe implementation here (which
is most games).