Implement scalable data generation
Fix the issues with hard-coded module paths and failing data
serialization in the previous commit so that users may specify any
module path they like as the simulator and run it on as many threads as
they choose.

The consumer runs separately and continually updates the model with the
latest data.